requirements.txt
python-dateutil >= 2.3
mock==1.0.1